Move UserBuilder to opengever.testing.
